Moved ReplicationState to a new file; combined epoch and epoch_history
