+++ src/storage/v2/replication/replication.hpp
@@ -1,0 +1,124 @@
+// Copyright 2023 Memgraph Ltd.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
+// License, and you may not use this file except in compliance with the Business Source License.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+#pragma once
+
+#include "storage/v2/inmemory/label_index.hpp"
+#include "storage/v2/inmemory/label_property_index.hpp"
+#include "storage/v2/storage.hpp"
+
+/// REPLICATION ///
+#include "rpc/server.hpp"
+#include "storage/v2/replication/config.hpp"
+#include "storage/v2/replication/enums.hpp"
+#include "storage/v2/replication/replication_persistence_helper.hpp"
+#include "storage/v2/replication/rpc.hpp"
+#include "storage/v2/replication/serialization.hpp"
+
+#include "storage/v2/replication/global.hpp"
+
+// TODO use replication namespace
+namespace memgraph::storage {
+
+class InMemoryStorage;
+class ReplicationServer;
+class ReplicationClient;
+
+// TODO: decouple from InMemoryStorage
+struct ReplicationState {
+  enum class RegisterReplicaError : uint8_t {
+    NAME_EXISTS,
+    END_POINT_EXISTS,
+    CONNECTION_FAILED,
+    COULD_NOT_BE_PERSISTED
+  };
+
+  // TODO: This mirrors the logic in InMemoryConstructor; make it independent
+  ReplicationState(bool restore, std::filesystem::path durability_dir);
+
+  // Generic API
+  void Reset();
+  // TODO: Just check if server exists -> you are REPLICA
+  replication::ReplicationRole GetRole() const { return replication_role_.load(); }
+
+  bool SetMainReplicationRole(Storage *storage);  // Set the instance to MAIN
+  // TODO: ReplicationServer/Client uses InMemoryStorage* for RPC callbacks
+  bool SetReplicaRole(io::network::Endpoint endpoint, const replication::ReplicationServerConfig &config,
+                      InMemoryStorage *storage);  // Sets the instance to REPLICA
+  // Generic restoration
+  void RestoreReplicationRole(InMemoryStorage *storage);
+
+  // MAIN actually doing the replication
+  bool AppendToWalDataDefinition(uint64_t seq_num, durability::StorageGlobalOperation operation, LabelId label,
+                                 const std::set<PropertyId> &properties, uint64_t final_commit_timestamp);
+  void InitializeTransaction(uint64_t seq_num);
+  void AppendDelta(const Delta &delta, const Vertex &parent, uint64_t timestamp);
+  void AppendDelta(const Delta &delta, const Edge &parent, uint64_t timestamp);
+  bool FinalizeTransaction(uint64_t timestamp);
+
+  // MAIN connecting to replicas
+  utils::BasicResult<RegisterReplicaError> RegisterReplica(std::string name, io::network::Endpoint endpoint,
+                                                           replication::ReplicationMode replication_mode,
+                                                           replication::RegistrationMode registration_mode,
+                                                           const replication::ReplicationClientConfig &config,
+                                                           InMemoryStorage *storage);
+  bool UnregisterReplica(std::string_view name);
+
+  // MAIN reconnecting to replicas
+  void RestoreReplicas(InMemoryStorage *storage);
+
+  // MAIN getting info from replicas
+  // TODO make into const (problem with SpinLock and WithReadLock)
+  std::optional<replication::ReplicaState> GetReplicaState(std::string_view name);
+  std::vector<ReplicaInfo> ReplicasInfo();
+
+  const auto &GetEpoch() const { return epoch_; }
+
+  // MAIN
+  void NewEpoch(uint64_t timestamp);
+  // REPLICA
+  void AppendEpoch(std::string new_epoch, uint64_t timestamp) {
+    epoch_.history.emplace_back(std::exchange(epoch_.id, new_epoch), timestamp);
+  }
+  void SetEpoch(std::string new_epoch) {}
+  auto &GetEpoch() { return epoch_; }
+
+ private:
+  bool ShouldStoreAndRestoreReplicationState() const { return nullptr != durability_; }
+
+  void SetRole(replication::ReplicationRole role) { return replication_role_.store(role); }
+
+  // NOTE: Server is not in MAIN it is in REPLICA
+  std::unique_ptr<ReplicationServer> replication_server_{nullptr};
+
+  // We create ReplicationClient using unique_ptr so we can move
+  // newly created client into the vector.
+  // We cannot move the client directly because it contains ThreadPool
+  // which cannot be moved. Also, the move is necessary because
+  // we don't want to create the client directly inside the vector
+  // because that would require the lock on the list putting all
+  // commits (they iterate list of clients) to halt.
+  // This way we can initialize client in main thread which means
+  // that we can immediately notify the user if the initialization
+  // failed.
+  using ReplicationClientList = utils::Synchronized<std::vector<std::unique_ptr<ReplicationClient>>, utils::SpinLock>;
+  ReplicationClientList replication_clients_;
+
+  std::atomic<replication::ReplicationRole> replication_role_{replication::ReplicationRole::MAIN};
+
+  std::unique_ptr<kvstore::KVStore> durability_;
+
+  ReplicationEpoch epoch_;
+
+  mutable std::mutex mtx_;  // TODO: Probably not needed
+};
+
+}  // namespace memgraph::storage
