+++ src/storage/v2/replication/replication_server.hpp
@@ -13,10 +13,12 @@
 
 #include "rpc/server.hpp"
 #include "slk/streams.hpp"
-#include "storage/v2/inmemory/storage.hpp"
+#include "storage/v2/replication/global.hpp"
 #include "storage/v2/replication/replication_client.hpp"
 
 namespace memgraph::storage {
+
+class InMemoryStorage;
 
 class ReplicationServer {
  public:
